Revert "Version 1.0.0" (#113)
Reverts splunk/addonfactory-ucc-test#111

+++ splunk_add_on_ucc_modinput_test/typing.py
@@ -15,7 +15,7 @@
 #
 from typing import Any, Callable, Dict, Generator, Tuple, Union, Optional
 
-ProbeGenType = Generator[float, None, Optional[bool]]
+ProbeGenType = Generator[int, None, Optional[bool]]
 ProbeGenFnType = Callable[..., ProbeGenType]
 ProbeRegularFnType = Callable[..., Any]
 ProbeFnType = Union[ProbeGenFnType, ProbeRegularFnType]
